Refactor: error boundary 리팩터링
