+++ src/app/id-card/[id]/components/CommentCount/index.ts
