+++ src/app/id-card/[id]/components/IdCardDetail/index.ts
@@ -1,0 +1,1 @@
+export * from './IdCardDetail';
